Fix failed AsyncioTest.test_send_failure and clean up tests
