+++ tests/run-unit-tests.sh
@@ -24,6 +24,9 @@ ROOT_DIR=$(git rev-parse --show-toplevel)
 cd $ROOT_DIR/tests
 
 python3 custom_logger_test.py
+python3 debug_logger_test.py
 python3 interrupted_test.py
 python3 pulsar_test.py
+python3 schema_test.py
+python3 reader_test.py
 python3 asyncio_test.py
